First minimal contract structure
The contract implements a way to purchase the royalty by sending the
minimum amount requested by the contract creator and also a function
to check wether an account has already purchased the royalty.

The data structure adopted to keep a record of accounts who have already
payed accomodates the fact that in Solidity 0.4.21 mappings do not allow
to check whether a key already exists. In a mapping all possible keys
exist and all values are initialized to 0.
In order to do check if a user has already payed a supporting array of
addresses is needed. The mapping is then used to map an address to an
index in the aforementioned array.